Feat: RecordTripCard.vue -> Trip 타입 파일 trip.entity.ts 로 이동
