--- src/shared/types/trip.ts
@@ -1,10 +1,0 @@
-// src/shared/types/trip.ts
-
-export interface Trip {
-  tripId: number
-  tripName: string
-  status: string
-  startDate: string
-  endDate: string
-  locationName: string
-}
